Update frequency from 2Mhz to 1Mhz
Apparently there was a mismatch between the GNU Radio recording script,
and the scripts we were given.

Also fixed a bug that had led me to set the limit value for the time
difference between two packets to be way higher than it should have
been...turned out I wasn't updating the last time-index we had seen
in the event we saw a repeated packet.  Oops!