Updates to support long-presses better
Given that I have steps #3 and #4 separated, we end up losing too much
data in the pipeline to be able to support both rapid-fire detection,
and long-press detection.  As such, I'm going for long-presses.  Roll
back this commit if you want to see rapid-fire work.  And maybe I'll
fix this...after the assignment is due, of course.  This is already
beyond what as asked for, don't need to spend further time on it. :)