Requisite big ol' pile of fixes
As often follows big ol' piles of updates, here is a big ol' pile of
fixes.

- newBeatData takes IntBuffer, instead of array.  Apparently creating it
as an array just throws exceptions.  Didn't care to figure out why. :)
- Miscellaneous logs throughout either commented out or added
- Updated Pulseduino to handle median counting better (more like A2).
Also added a 200ms wait between beats.  That's 300 BPM, and since my
device is not going to call 911 for you, I'm taking the hear-no-evil
approach.